#142: update test coverage
